+++ catscripts/Government/Countries/Bahamas/Bahamas.sh
@@ -1,14 +1,31 @@
 #!/bin/bash
 
-egrep -i 'Bahamas' newpages.txt >> Bahamas.txt
+KEYWORDS_BAHAMAS="Bahamas"
+KEYWORDS_BAHAMAS_ALL="$KEYWORDS_BAHAMAS"
 
-BAHAMAS=`stat --print=%s Bahamas.txt`
-
-if [ $BAHAMAS -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Bahamas.txt"
-  export CATNAME="The Bahamas"
-  $CATEGORIZE
-fi
 
-rm Bahamas.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Bahamas\n"
+  fi
+
+  BAHAMAS="$(egrep -i "$KEYWORDS_BAHAMAS" newpages.txt)"
+
+  if [ "$BAHAMAS" != "" ];
+  then
+    printf "%s" "$BAHAMAS" > Bahamas.txt
+    export CATFILE="Bahamas.txt"
+    export CATNAME="The Bahamas"
+    $CATEGORIZE
+    rm Bahamas.txt
+    unset BAHAMAS
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Bahamas\n"
+  fi
+
+fi
